sec lib refactoring
