--- api-security/src/main/java/gov/ca/cwds/security/annotations/AclAuthorize.java
@@ -1,7 +1,0 @@
-package gov.ca.cwds.security.annotations;
-
-/**
- * Created by dmitry.rudenko on 9/27/2017.
- */
-public @interface AclAuthorize {
-}
