+++ api-security/src/test/resources/shiro.ini
@@ -7,7 +7,7 @@ securityManager.authorizer.permissionResolver = $globalPermissionResolver
 # -------------
 # Perry Realm
 # -------------
-perryRealm = gov.ca.cwds.security.realm.AbacRealm
+perryRealm = gov.ca.cwds.security.realm.TestRealm
 perryRealm.keyStorePath=config/enc.jceks
 perryRealm.keyStoreAlias=test
 perryRealm.keyStorePassword=test
